+++ NEWS.md
@@ -1,3 +1,10 @@
+## R package processR version 0.2.4.1
+=====================================
+(2020-Feb-23)
+
+- Make modelsSummaryTable function compatible with flextable(>= 0.5.8)
+
+
 ## R package processR version 0.2.4
 ===================================
 (2020-Feb-22)